refactor: ファイルごとの関数配置を変更。ファイル名を変更
